detector/requirements.txt
Flask==2.0.1
pika==1.2.0
Pillow~=9.5.0
requests==2.26.0
Werkzeug==2.0.1
tenacity==8.0.1
opencv-python-headless~=4.7.0.72
# opencv-python~=4.7.0.72
ultralytics~=8.0.118
numpy~=1.24.3
matplotlib~=3.7.1
pandas~=1.5.3
dill~=0.3.6
py-cpuinfo~=9.0.0
yadisk~=1.3.3